request validation and generateUploadUrl - both untested

+++ course-04/project/c4-final-project-starter-code/backend/src/lambda/http/deleteTodo.ts
@@ -1,6 +1,7 @@
 import 'source-map-support/register'
 import { createLogger } from '../../utils/logger'
 import { deleteTodo } from '../../businessLogic/todos'
+import { getUserId } from '../utils'
 
 const logger = createLogger('deleteTodo')
 
@@ -9,10 +10,22 @@ import { APIGatewayProxyEvent, APIGatewayProxyResult, APIGatewayProxyHandler } f
 export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   logger.info('processing event', {key: event})
   const todoId = event.pathParameters.todoId
-  const userId = "1"
+  const userId = getUserId(event)
+
+  if (!userId) {
+    return {
+      statusCode: 401,
+      headers: {
+        'Access-Control-Allow-Origin': '*'
+      },
+      body: JSON.stringify({
+        error: 'User is not authorized'
+      })
+    }
+  }
 
   const result  = await deleteTodo(userId, todoId)
-  
+
   if(result == undefined) {
     logger.warn('tried to delete a non-existing item', {key: result})
     return {
